collision

+++ src/main/java/me/meczka/managers/GameCalcuator.java
@@ -4,7 +4,7 @@ package me.meczka.managers;
  * Created by Chilik on 25.01.2018.
  */
 public class GameCalcuator {
-    public static int pixelsToTiles(int pixels)
+    public static float pixelsToTiles(int pixels)
     {
         return pixels/50;
     }
